chore: fix horizontal scroll breaking w/long titles on home page
With long unbroken strings in a title on the home page the
mobile page has horizontal scroll where it shouldn't and
has a bad user experiance. This should fix it so it doesn't
continue to force horizontal scroll.

This prefers text-wrap: pretty, but falls back to the more
broadly adopted text-wrap: balance.